Tests/LibWeb/Text/expected/shared-array-buffer-shared-across-message-channel.txt

port2 saw initial: 1
after port2 writes 2, port1 view sees: 2
